# Import data using ImportWithSize call, not manual resizing