--- a/qubesadmin/tests/utils.py
+++ b/qubesadmin/tests/utils.py
@@ -1,4 +1,4 @@
-# -*- encoding: utf8 -*-
+# -*- encoding: utf-8 -*-
 #
 # The Qubes OS Project, http://www.qubes-os.org
 #