See the API call in QubesOS/qubes-core-admin#309.