@pwmarcz 2020-01-20
Some benchmarks. I tried importing a 2GB image for a VM (non-sparse, containing a file full of random data). The command was `time qvm-volume import test-import:private data`.

The real time is a bit noisy but user/sys are more consistent:

|version|real|user|sys|
|-|-|-|-|
|current|49.3|1.9|6.2|
|current|62.0|2.0|5.2|
|new (default buf size, 64K)|57.9|5.6|13.5|
|new (default buf size, 64K)|54.3|5.7|14.2|
|new (buf size 1M)|55.0|2.3|9.4|
|new (buf size 1M)|70.6|2.3|10.0|