@pwmarcz 2020-01-20
Here are the numbers when running from a VM:

|version|real|user|sys|
|-|-|-|-|
|current|105.5|2.7|7.1|
|current|86.9|2.7|6.5|
|new (default buf size, 64K)|80.7|5.1|11.9|
|new (default buf size, 64K)|127.4|5.0|12.1|
|new (buf size 1M)|98.1|3.6|11.3|
|new (buf size 1M)|95.5|3.7|12.6|