@marmarek 2020-01-20
Since multiple processes, some in different VM are involved user and sys times are meaningless for overall performance (more important is the actual throughput, not CPU time used to achieve it).
Can you try to get more consistent results (real time)? IDK, maybe clear disk caches before the call (`echo 3 > /proc/sys/vm/drop_caches`)?