@pwmarcz 2020-01-20
Here are the results for 3 runs each, from a VM, dropping the cache just before the import, and a relatively quiet system (I did not do anything on the machine during):

|version|real|
|current|80|
|current|94|
|current|100|
|new (64K)|86|
|new (64K)|108|
|new (64K)|102|
|new (1M)|70|
|new (1M)|106|
|new (1M)|103|