@pwmarcz 2020-01-20
And here is what happened when I tried it in dom0 and even shut down noisy programs like Firefox:

|version|real|
|-|-|
|current|63|
|current|102|
|current|103|
|new (64K)|60|
|new (64K)|79|
|new (64K)|97|
|new (1M)|66|
|new (1M)|93|
|new (1M)|86|