@marmarek 2020-01-20
It's interesting that the first call consistently is faster.
Anyway, 1) slowdown if any isn't big b) I don't see conclusive difference between 64K and 1M buffer, so I'd go with the default.